requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
easydict==1.9
et-xmlfile==1.1.0
gitdb==4.0.7
GitPython==3.1.18
-e git+https://github.com/jacobgil/pytorch-grad-cam.git@c009adac84ec83565c88c692a13bd3581e7c24eb#egg=grad_cam
idna==2.10
importlib-resources==5.2.0
iterative-stratification==0.1.6
itsdangerous==2.0.1
joblib==1.0.1
kiwisolver==1.3.1
kornia==0.5.5
matplotlib==3.4.2
numpy==1.19.0
opencv-python==4.5.2.54
openpyxl==3.0.7
pandas==1.3.0
pathtools==0.1.2
Pillow==8.2.0
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyaml==20.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-cache==0.6.4
scikit-learn==0.24.2
scipy==1.7.0
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
threadpoolctl==2.1.0
-e git+https://github.com/rwightman/pytorch-image-models.git@79190534256fef6b665da20bd30f9d3fb194daeb#egg=timm
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
ttach==0.0.3
typing-extensions==3.10.0.0
url-normalize==1.4.3
urllib3==1.26.6
zipp==3.5.0
